+++ src/templates/mod.rs
@@ -14,18 +14,11 @@ pub use self::utils::*;
 #[cfg(feature = "mime03")]
 use mime::Mime;
 
-#[cfg(feature = "mime02")]
-/// Documentation mock.  The real Mime type comes from the `mime` crate.
-pub type Mime = u8; // mock
-
 /// A static file has a name (so its url can be recognized) and the
 /// actual file contents.
 ///
 /// The content-type (mime type) of the file is available as a
-/// static field when building ructe with the `mime03` feature or
-/// as the return value of a method when building ructe with the
-/// `mime02` feature (in `mime` version 0.2.x, a Mime cannot be
-/// defined as a part of a const static value.
+/// static field when building ructe with the `mime03` feature.
 pub struct StaticFile {
     /// The actual static file contents.
     pub content: &'static [u8],
@@ -37,19 +30,6 @@ pub struct StaticFile {
     /// crate version 0.3.x.
     #[cfg(feature = "mime03")]
     pub mime: &'static Mime,
-}
-
-impl StaticFile {
-    /// Get the mime type of this static file.
-    ///
-    /// Currently, this method parses a (static) string every time.
-    /// A future release of `mime` may support statically created
-    /// `Mime` structs, which will make this nicer.
-    #[allow(unused)]
-    #[cfg(feature = "mime02")]
-    pub fn mime(&self) -> Mime {
-        unimplemented!()
-    }
 }
 
 #[test]